memcached/memcached-server/cache server/app/search/requirements.txt
python-memcached
uhashring==2.3
numpy==1.21.5